Redirect legacy glean dictionary URLs
For glean-dictionary.netlify.app:

Use netlify's redirect to direct glean-dictionary.netlify.app => dictionary.telemetry.mozilla.org

For protosaur, we need to resort to a javascript-based solution.

This change is *not* meant to be merged into main. It's a bespoke solution
and we can remove it once we're pretty sure people aren't using the old
domains.